Share biome resource configurations wherever possible.
Using `BiomeExtends` with `ResourceInheritance: true` allows biomes to
inherit their resource configuration from another biome. This minimises
the effort in making a change to resources that affects all biomes, such
as changing the quartz distribution. It also minimises the chances of
forgetting to change a biome where it is required.

However, OTG's resource inheritance mechanism is idiosyncratic, and
poorly documented. Testing shows that if a resource is added to a biome
that inherits the same type of resource from another biome, the inherited
configuration is ignored. The new resource definition is treated as
completely overriding an inherited resource of the same block type
rather than adding to it.

In the case of the nether biomes, I want to prevent resources like quartz
ore and lava lakes from spawning at the surface in biomes with surface
concrete powder and sand. Even though such deposits may be restricted to
spawning in netherrack only, their size makes them show through. To solve
that, I need those resources to only spawn at Y60 or below. However, in
biomes that are not overlaid with non-netherrack blocks, I want quartz to
spawn higher than Y60 as well.

Due to the aforementioned limitations of the resource inheritance mechanism,
I cannot share the Y0-60 quartz etc configuration between, say, OchreDesert
and more vanilla-like nether biomes, because I would need to override the
quartz distribution in all of the biomes where it occurs at higher Y.

So, for the purposes of minimising duplicated configuration, I distinguish
two categories of biomes and inherit their resource configurations from two
distinct custom biomes:

 * Biomes that only have a netherrack surface inherit their resources from
   `NetherrackOnlyResources.bc`.
 * Biomes that have something non-vanilla overlaying netherrack at the
   surface inherit resources from `OverlaidNetherrackResources.bc`.

The world is configured not to spawn either of these custom biomes in the
NormalBiomes group. They are simply containers for the resource
configurations. There is still some duplication of resource configuration,
but it is limited to just those two files.

Changes:

 * Add non-spawned custom biomes `NetherrackOnlyResources.bc` and
   `OverlaidNetherrackResources.bc` and configure all spawned biomes to
   inherit resources from one or the other.
 * Increase cave frequency.
 * Set a lava layer just above bedrock in all biomes.
 * Git ignore `*.bc.inherited` files generated by OTG when reading
   biomes with inherited resources.